requirements.txt
streamlit
chromadb
requests
pypdf
sentence-transformers>=2.3.0
torch
transformers>=4.30.0

# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
